Store_19 - feat: add settings data sort
