+++ src/data/data.json
@@ -5,7 +5,7 @@
       "title": "KONGSBERG",
       "description": "Fabric upholstery and wood. Accommodates spring and springless mattresses of size 140x200 cm. Without slatted base and mattress. 156x218 cm, height 101 cm",
       "price": 250,
-      "discountPercentage": 15,
+      "discount": 15,
       "rating": 4.5,
       "stock": 94,
       "type": "Bed",
@@ -23,7 +23,7 @@
       "title": "ABILDRO",
       "description": "Metal. Accommodates box mattress, spring and springless mattresses size 140x200 cm. Without base and mattress. 147x217 cm, height 100 cm",
       "price": 149,
-      "discountPercentage": 10,
+      "discount": 10,
       "rating": 3.9,
       "stock": 34,
       "type": "Bed",
@@ -41,7 +41,7 @@
       "title": "SALLINGE",
       "description": "Spruce massif. Accommodates box mattress, spring and springless mattresses size 140x200 cm. Without base and mattress. 148x208 cm, height 50 cm",
       "price": 110,
-      "discountPercentage": 20,
+      "discount": 20,
       "rating": 4.1,
       "stock": 56,
       "type": "Bed",
@@ -59,7 +59,7 @@
       "title": "TAPDRUP",
       "description": "Deco veneer. 3 drawer chest oak. W71 x H69 x D35 cm",
       "price": 50,
-      "discountPercentage": 7,
+      "discount": 7,
       "rating": 4.2,
       "stock": 123,
       "type": "Chest of Drawer",
@@ -77,7 +77,7 @@
       "title": "BJERRINGBRO",
       "description": "6 drawer chest. Natural/white. Solid wood and MDF. W83 x H80 x D38 cm",
       "price": 99,
-      "discountPercentage": 12,
+      "discount": 12,
       "rating": 3.1,
       "stock": 32,
       "type": "Chest of Drawers",
@@ -94,7 +94,7 @@
       "title": "GEDVED",
       "description": "3-seater sofa in fabric, grey. Seat cushion with pocket springs and foam padding. Back cushion in foam. Legs in solid wood. W210 x H85 x D84 cm",
       "price": 490,
-      "discountPercentage": 20,
+      "discount": 20,
       "rating": 4.7,
       "stock": 83,
       "type": "Sofa",
@@ -112,7 +112,7 @@
       "title": "EGEDAL",
       "description": "2.5-seater sofa in fabric, grey. Seat and back cushions in foam. Legs in solid wood. W170 x H82 x D81 cm",
       "price": 249,
-      "discountPercentage": 21,
+      "discount": 21,
       "rating": 4.1,
       "stock": 50,
       "type": "Sofa",
@@ -130,7 +130,7 @@
       "title": "KONGSMARK",
       "description": "2.5-seater sofa in fabric, dark grey. Seat and back cushions in foam. Legs in solid wood. W209 x H80 x D91 cm",
       "price": 525,
-      "discountPercentage": 22,
+      "discount": 22,
       "rating": 4.3,
       "stock": 68,
       "type": "Sofa",
@@ -148,7 +148,7 @@
       "title": "INBOOK",
       "description": "Bamboo. D60 x H45 cm.",
       "price": 55,
-      "discountPercentage": 10,
+      "discount": 10,
       "rating": 4.7,
       "stock": 96,
       "type": "Table",
@@ -166,7 +166,7 @@
       "title": "FALSLED",
       "description": "Nest of tables. Oak veneer and tempered glass. D50/40 x H45/40 cm",
       "price": 79,
-      "discountPercentage": 6,
+      "discount": 6,
       "rating": 5,
       "stock": 89,
       "type": "Table",
@@ -184,7 +184,7 @@
       "title": "SKIBET",
       "description": "Dining table. Solid oak and light oak veneer. D120 x H75 cm.",
       "price": 350,
-      "discountPercentage": 11,
+      "discount": 11,
       "rating": 4.2,
       "stock": 65,
       "type": "Table",
@@ -202,7 +202,7 @@
       "title": "EGENS",
       "description": "Dining table. Solid oak and black oak veneer 90x190/270",
       "price": 325,
-      "discountPercentage": 26,
+      "discount": 26,
       "rating": 4.6,
       "stock": 52,
       "type": "Table",
@@ -220,7 +220,7 @@
       "title": "HVIDOVRE",
       "description": "Dining chair. Fabric, oak veneer and steel.",
       "price": 85,
-      "discountPercentage": 5,
+      "discount": 5,
       "rating": 5.0,
       "stock": 68,
       "type": "Chair",
@@ -238,7 +238,7 @@
       "title": "JELS",
       "description": "Dining chair. Pine massif, oak. 42x50 cm, height 86 cm",
       "price": 80,
-      "discountPercentage": 20,
+      "discount": 20,
       "rating": 4.3,
       "stock": 114,
       "type": "Chair",
@@ -256,7 +256,7 @@
       "title": "TJELE",
       "description": "Office chair. Faux leather. With adjustable height and stepless tilt mechanism. W68 x H115-125 x D77 cm.",
       "price": 115,
-      "discountPercentage": 13,
+      "discount": 13,
       "rating": 4.0,
       "stock": 105,
       "type": "Chair",
@@ -274,7 +274,7 @@
       "title": "ASPERUP",
       "description": "Office chair. Fabric and mesh. Base in steel. With lumbar support, flip-up armrests, adjustable height and stepless tilt mechanism. W59 x H95-106 x D61 cm.",
       "price": 85,
-      "discountPercentage": 23,
+      "discount": 23,
       "rating": 3.0,
       "stock": 85,
       "type": "Chair",
@@ -292,7 +292,7 @@
       "title": " TRAPPEDAL",
       "description": "Shelving unit. Deco veneer and metal. 5 shelves. W70 x H109 x D30 cm.",
       "price": 70,
-      "discountPercentage": 22,
+      "discount": 22,
       "rating": 4.1,
       "stock": 78,
       "type": "Bkcases/Room divider",
@@ -310,7 +310,7 @@
       "title": "MARKSKEL",
       "description": "Bookcase. Deco veneer and MDF. Drawers with full extension slides. W90 x H153 x D36 cm.",
       "price": 165,
-      "discountPercentage": 30,
+      "discount": 30,
       "rating": 4.5,
       "stock": 88,
       "type": "Bkcase/Room divider",
@@ -328,7 +328,7 @@
       "title": "JENSLEV",
       "description": "Deco veneer. 3 doors + 3 drawers. Wardrobe interior: 4 shelves and 2 hanging rails. W145 x H176 x D50 cm.",
       "price": 230,
-      "discountPercentage": 15,
+      "discount": 15,
       "rating": 4.1,
       "stock": 54,
       "type": "Wardrobe",
@@ -346,7 +346,7 @@
       "title": "FREDENSBORG",
       "description": "Deco veneer. 4 doors + 2 drawers. Wardrobe interior: 8 shelves and 1 hanging rail. W183 x H200 x D64 cm.",
       "price": 425,
-      "discountPercentage": 20,
+      "discount": 20,
       "rating": 4.4,
       "stock": 40,
       "type": "Wardrobe",
